modified to use hybrid algorithm service
